fixed interface DiscountStrat

+++ src/practicetest/DiscountStrat.java
@@ -11,6 +11,6 @@ package practicetest;
  * @author khalloway
  */
 public interface DiscountStrat {
-    public abstract double getDiscountedPrice(double price, int quantity);
-    public abstract double getAmountSaved(double price, int quantity);
+    public abstract double getDiscountedPrice(double price, double quantity);
+    public abstract double getAmountSaved(double price, double quantity);
 }
