+++ src/practicetest/Product.java
@@ -55,11 +55,11 @@ public class Product {
         this.price = price;
     }
     
-    public double getProdDiscountedPrice(int quantity){
+    public double getProdDiscountedPrice(double quantity){
         return strat.getDiscountedPrice(price, quantity);
     }
     
-    public double getProdAmountSaved(int quantity){
+    public double getProdAmountSaved(double quantity){
         return strat.getAmountSaved(price, quantity);
     }
 }
